requirements.txt
streamlit
pandas
numpy
matplotlib
seaborn
scipy
tensorflow-cpu
plotly
pickle-mixin
sklearn